+++ org.eclipse.nebula.widgets.nattable.core/src/org/eclipse/nebula/widgets/nattable/conflation/EventConflaterChain.java
@@ -26,8 +26,8 @@ import org.eclipse.nebula.widgets.nattable.util.Scheduler;
  */
 public class EventConflaterChain implements IEventConflater {
 
-    public static final int DEFAULT_INITIAL_DELAY = 100;
-    public static final int DEFAULT_REFRESH_INTERVAL = 20;
+    public static final long DEFAULT_INITIAL_DELAY = 100;
+    public static final long DEFAULT_REFRESH_INTERVAL = 20;
     private static final Scheduler scheduler = new Scheduler("EventConflaterChain"); //$NON-NLS-1$
 
     private final List<IEventConflater> chain = new LinkedList<IEventConflater>();
@@ -40,7 +40,10 @@ public class EventConflaterChain implements IEventConflater {
         this(DEFAULT_REFRESH_INTERVAL, DEFAULT_INITIAL_DELAY);
     }
 
-    public EventConflaterChain(int refreshInterval, int initialDelay) {
+    /**
+     * @since 2.0
+     */
+    public EventConflaterChain(long refreshInterval, long initialDelay) {
         this.refreshInterval = refreshInterval;
         this.initialDelay = initialDelay;
     }
